+++ intelmq-manager/php/config.php
@@ -13,4 +13,5 @@
 
     $BOT_CONFIGS_REJECT_REGEX = '/[^[:print:]\n\r\t]/';
     $BOT_ID_REJECT_REGEX = '/[^A-Za-z0-9.-]/';
+    $VERSION = "0.2.2";
 ?>